# trace-core: Dispatchers unset themselves