This branch changes `dispatcher::get_default` to unset the thread's
current dispatcher while the reference to it is held by the closure.
This prevents infinite loops if the subscriber calls code paths which
emit events or construct spans. 

Note that this also means that nested calls to `get_default` inside of a
`get_default` closure will recieve a `None` dispatcher rather than the
"actual" dispatcher. However, it was necessary to unset the default in
`get_default` rather than in dispatch methods such as `Dispatch::enter`,
as when those functions are called, the current state has already been
borrowed.

This introduces a little more overhead into span creation and events.

Before:
```

     Running target/release/deps/no_subscriber-09f0b39a00181a97

running 5 tests
test bench_1_atomic_load              ... bench:           0 ns/iter (+/- 0)
test bench_costly_field_no_subscriber ... bench:           2 ns/iter (+/- 0)
test bench_log_no_logger              ... bench:           0 ns/iter (+/- 0)
test bench_no_span_no_subscriber      ... bench:           0 ns/iter (+/- 0)
test bench_span_no_subscriber         ... bench:           2 ns/iter (+/- 0)

test result: ok. 0 passed; 0 failed; 0 ignored; 5 measured; 0 filtered out

     Running target/release/deps/subscriber-992535c5d1c348ad

running 5 tests
test enter_span              ... bench:           3 ns/iter (+/- 0)
test span_no_fields          ... bench:          38 ns/iter (+/- 8)
test span_repeatedly         ... bench:       4,560 ns/iter (+/- 663)
test span_with_fields        ... bench:          49 ns/iter (+/- 17)
test span_with_fields_record ... bench:         363 ns/iter (+/- 71)
```

After:
```
     Running target/release/deps/no_subscriber-09f0b39a00181a97

running 5 tests
test bench_1_atomic_load              ... bench:           0 ns/iter (+/- 0)
test bench_costly_field_no_subscriber ... bench:           2 ns/iter (+/- 0)
test bench_log_no_logger              ... bench:           0 ns/iter (+/- 0)
test bench_no_span_no_subscriber      ... bench:           0 ns/iter (+/- 0)
test bench_span_no_subscriber         ... bench:           2 ns/iter (+/- 0)

test result: ok. 0 passed; 0 failed; 0 ignored; 5 measured; 0 filtered out

     Running target/release/deps/subscriber-992535c5d1c348ad

running 5 tests
test enter_span              ... bench:           3 ns/iter (+/- 0)
test span_no_fields          ... bench:          48 ns/iter (+/- 9)
test span_repeatedly         ... bench:       5,274 ns/iter (+/- 596)
test span_with_fields        ... bench:          55 ns/iter (+/- 11)
test span_with_fields_record ... bench:         346 ns/iter (+/- 71)
```

Closes #1032 

Signed-off-by: Eliza Weisman <eliza@buoyant.io>