@carllerche 2019-04-04
Looking good :+1: I only skimmed it so far. I will review it in closer detail tomorrow.

Have you considered flagging the dispatcher as entered instead of swapping out then back? I wonder if that would make an impact in the benches you run.

Roughly, `Dispatcher` would have a bool `entered`. If that flag is set, then it does not forward any events to the subscriber.