@hawkw 2019-04-04
@carllerche We still need to store that bool as part of the thread-local state, rather than on `Dispatch`, since `Dispatch` needs to be `Sync` (precluding the use of a `Cell`) as it's a field on `Span` (which we are now asserting is `Sync`). 

Using a flag stored in the thread-local state rather than replacing the dispatcher doesn't seem to result in a big performance improvement:
```

running 5 tests
test enter_span              ... bench:           3 ns/iter (+/- 0)
test span_no_fields          ... bench:          47 ns/iter (+/- 6)
test span_repeatedly         ... bench:       5,063 ns/iter (+/- 957)
test span_with_fields        ... bench:          57 ns/iter (+/- 6)
test span_with_fields_record ... bench:         361 ns/iter (+/- 37)
```
However, it _does_ mean that calls to `dispatcher::get_default` can be nested, as long as  calls to the functions on `Dispatch` that need to be guarded are not also nested. This at least lets us avoid a potentially surprising edge case...