@carllerche 2019-04-09
@hawkw looking at this, the version w/ the bool flag needs to add a lot of conditionals where the original one did not... I do wonder what made the original (`Cell` and swap) slower.