@hawkw 2019-04-10
@carllerche personally, I would prefer to move forward with the bool flag rather than `Cell` and swap, because it means that nested calls to `get_default` don't behave surprisingly --- in the bool flag approach, the dispatcher is only disabled when we're actually inside it.