@hawkw 2019-04-15
@carllerche commit b0c6b3431e81763535579c51a5de0af51cf04c32 moves the check back into `dispatcher::get_default`, rather than guarding individual functions on the dispatcher.

This means that we don't have to access the TLS var twice, and we don't have to access it at all when entering a span handle that already exists. This pretty much entirely negates the performance hit caused by adding the check.