@hawkw 2019-04-15
Looks like the build failure is due to Azure Pipelines [flaking out again](https://dev.azure.com/tokio-rs/Tokio/_build/results?buildId=438&view=results);  I restarted it.