@hawkw 2019-04-15
FreeBSD CI is also failing due to an unrelated (and possibly transient?) failure in the `threadpool` tests:
```
failures:
---- blocking_thread_does_not_take_over_shutdown_worker_thread stdout ----
thread 'blocking_thread_does_not_take_over_shutdown_worker_thread' panicked at 'called `Result::unwrap()` on an `Err` value: RecvError', src/libcore/result.rs:997:5
failures:
    blocking_thread_does_not_take_over_shutdown_worker_thread
test result: FAILED. 6 passed; 1 failed; 0 ignored; 0 measured; 0 filtered out
```
I restarted it.